# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "TÖPP API",
    "description": "REST API интернет-магазина ковров; цены в целых рублях. /docs для интерактивного OpenAPI.",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/products": {
      "get": {
        "summary": "List Products",
        "operationId": "list_products_api_v1_products_get",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Limit"
            }
          },
          {
            "name": "style",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Style"
            }
          },
          {
            "name": "colorCode",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Colorcode"
            }
          },
          {
            "name": "countryCode",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Countrycode"
            }
          },
          {
            "name": "size",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Size"
            }
          },
          {
            "name": "lengthCm",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer",
                  "exclusiveMinimum": 0
                },
                {
                  "type": "null"
                }
              ],
              "title": "Lengthcm"
            }
          },
          {
            "name": "widthCm",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer",
                  "exclusiveMinimum": 0
                },
                {
                  "type": "null"
                }
              ],
              "title": "Widthcm"
            }
          },
          {
            "name": "minPrice",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer",
                  "minimum": 0
                },
                {
                  "type": "null"
                }
              ],
              "title": "Minprice"
            }
          },
          {
            "name": "maxPrice",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer",
                  "minimum": 0
                },
                {
                  "type": "null"
                }
              ],
              "title": "Maxprice"
            }
          },
          {
            "name": "sort",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "pattern": "^(popular|newest|price_asc|price_desc|size|discount)$",
              "default": "newest",
              "title": "Sort"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          },
          "409": {
            "description": "Конфликт цен, остатков или условий",
            "content": {
              "application/json": {
                "example": {
                  "error": {
                    "code": "CART_CHANGED",
                    "message": "Цены или остатки изменились",
                    "details": {
                      "items": [
                        {
                          "variantId": "00000000-0000-4000-8000-000000000000",
                          "available": 0,
                          "expectedPrice": 12990,
                          "currentPrice": 13990
                        }
                      ]
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/products/{slug}": {
      "get": {
        "summary": "Product Detail",
        "operationId": "product_detail_api_v1_products__slug__get",
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Slug"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          },
          "409": {
            "description": "Конфликт цен, остатков или условий",
            "content": {
              "application/json": {
                "example": {
                  "error": {
                    "code": "CART_CHANGED",
                    "message": "Цены или остатки изменились",
                    "details": {
                      "items": [
                        {
                          "variantId": "00000000-0000-4000-8000-000000000000",
                          "available": 0,
                          "expectedPrice": 12990,
                          "currentPrice": 13990
                        }
                      ]
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/orders": {
      "post": {
        "summary": "Create Order",
        "operationId": "create_order_api_v1_orders_post",
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 8,
              "maxLength": 128,
              "title": "Idempotency-Key"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OrderIn"
              },
              "example": {
                "customerName": "Иван",
                "phone": "+79991234567",
                "address": {
                  "city": "Москва",
                  "street": "Лесная",
                  "house": "1"
                },
                "deliveryMethod": "manual_delivery",
                "paymentMethod": "on_delivery",
                "personalDataConsent": true
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "example": {
                  "orderId": "00000000-0000-4000-8000-000000000000",
                  "orderNumber": "f9126af6-bced-439a-a050-a77d9c1c570b",
                  "status": "new",
                  "currency": "RUB",
                  "subtotal": 12990,
                  "delivery": 0,
                  "total": 12990,
                  "deliveryMethod": "manual_delivery",
                  "paymentMethod": "on_delivery",
                  "items": [
                    {
                      "variantId": "00000000-0000-4000-8000-000000000000",
                      "name": "Лунна",
                      "sku": "SKU-001",
                      "sizeLabel": "160 × 230 см",
                      "unitPrice": 12990,
                      "quantity": 1,
                      "lineTotal": 12990
                    }
                  ]
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          },
          "409": {
            "description": "Конфликт цен, остатков или условий",
            "content": {
              "application/json": {
                "example": {
                  "error": {
                    "code": "CART_CHANGED",
                    "message": "Цены или остатки изменились",
                    "details": {
                      "items": [
                        {
                          "variantId": "00000000-0000-4000-8000-000000000000",
                          "available": 0,
                          "expectedPrice": 12990,
                          "currentPrice": 13990
                        }
                      ]
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "cartBearer": []
          }
        ]
      }
    },
    "/api/v1/admin/products": {
      "get": {
        "summary": "Products",
        "operationId": "products_api_v1_admin_products_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "409": {
            "description": "Конфликт цен, остатков или условий",
            "content": {
              "application/json": {
                "example": {
                  "error": {
                    "code": "CART_CHANGED",
                    "message": "Цены или остатки изменились",
                    "details": {
                      "items": [
                        {
                          "variantId": "00000000-0000-4000-8000-000000000000",
                          "available": 0,
                          "expectedPrice": 12990,
                          "currentPrice": 13990
                        }
                      ]
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "adminBasic": []
          }
        ]
      },
      "post": {
        "summary": "Product Create",
        "operationId": "product_create_api_v1_admin_products_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProductInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          },
          "409": {
            "description": "Конфликт цен, остатков или условий",
            "content": {
              "application/json": {
                "example": {
                  "error": {
                    "code": "CART_CHANGED",
                    "message": "Цены или остатки изменились",
                    "details": {
                      "items": [
                        {
                          "variantId": "00000000-0000-4000-8000-000000000000",
                          "available": 0,
                          "expectedPrice": 12990,
                          "currentPrice": 13990
                        }
                      ]
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "adminBasic": []
          }
        ]
      }
    },
    "/api/v1/admin/products/{id}": {
      "get": {
        "summary": "Product Detail",
        "operationId": "product_detail_api_v1_admin_products__id__get",
        "security": [
          {
            "adminBasic": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          },
          "409": {
            "description": "Конфликт цен, остатков или условий",
            "content": {
              "application/json": {
                "example": {
                  "error": {
                    "code": "CART_CHANGED",
                    "message": "Цены или остатки изменились",
                    "details": {
                      "items": [
                        {
                          "variantId": "00000000-0000-4000-8000-000000000000",
                          "available": 0,
                          "expectedPrice": 12990,
                          "currentPrice": 13990
                        }
                      ]
                    }
                  }
                }
              }
            }
          }
        }
      },
      "patch": {
        "summary": "Product Patch",
        "operationId": "product_patch_api_v1_admin_products__id__patch",
        "security": [
          {
            "adminBasic": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProductPatch"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          },
          "409": {
            "description": "Конфликт цен, остатков или условий",
            "content": {
              "application/json": {
                "example": {
                  "error": {
                    "code": "CART_CHANGED",
                    "message": "Цены или остатки изменились",
                    "details": {
                      "items": [
                        {
                          "variantId": "00000000-0000-4000-8000-000000000000",
                          "available": 0,
                          "expectedPrice": 12990,
                          "currentPrice": 13990
                        }
                      ]
                    }
                  }
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Product Delete",
        "operationId": "product_delete_api_v1_admin_products__id__delete",
        "security": [
          {
            "adminBasic": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          },
          "409": {
            "description": "Конфликт цен, остатков или условий",
            "content": {
              "application/json": {
                "example": {
                  "error": {
                    "code": "CART_CHANGED",
                    "message": "Цены или остатки изменились",
                    "details": {
                      "items": [
                        {
                          "variantId": "00000000-0000-4000-8000-000000000000",
                          "available": 0,
                          "expectedPrice": 12990,
                          "currentPrice": 13990
                        }
                      ]
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/products/{id}/variants": {
      "post": {
        "summary": "Variant Create",
        "operationId": "variant_create_api_v1_admin_products__id__variants_post",
        "security": [
          {
            "adminBasic": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VariantInput"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          },
          "409": {
            "description": "Конфликт цен, остатков или условий",
            "content": {
              "application/json": {
                "example": {
                  "error": {
                    "code": "CART_CHANGED",
                    "message": "Цены или остатки изменились",
                    "details": {
                      "items": [
                        {
                          "variantId": "00000000-0000-4000-8000-000000000000",
                          "available": 0,
                          "expectedPrice": 12990,
                          "currentPrice": 13990
                        }
                      ]
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/variants/{id}": {
      "patch": {
        "summary": "Variant Patch",
        "operationId": "variant_patch_api_v1_admin_variants__id__patch",
        "security": [
          {
            "adminBasic": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VariantPatch"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          },
          "409": {
            "description": "Конфликт цен, остатков или условий",
            "content": {
              "application/json": {
                "example": {
                  "error": {
                    "code": "CART_CHANGED",
                    "message": "Цены или остатки изменились",
                    "details": {
                      "items": [
                        {
                          "variantId": "00000000-0000-4000-8000-000000000000",
                          "available": 0,
                          "expectedPrice": 12990,
                          "currentPrice": 13990
                        }
                      ]
                    }
                  }
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Variant Delete",
        "operationId": "variant_delete_api_v1_admin_variants__id__delete",
        "security": [
          {
            "adminBasic": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          },
          "409": {
            "description": "Конфликт цен, остатков или условий",
            "content": {
              "application/json": {
                "example": {
                  "error": {
                    "code": "CART_CHANGED",
                    "message": "Цены или остатки изменились",
                    "details": {
                      "items": [
                        {
                          "variantId": "00000000-0000-4000-8000-000000000000",
                          "available": 0,
                          "expectedPrice": 12990,
                          "currentPrice": 13990
                        }
                      ]
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/orders": {
      "get": {
        "summary": "Orders",
        "operationId": "orders_api_v1_admin_orders_get",
        "security": [
          {
            "adminBasic": []
          }
        ],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          },
          "409": {
            "description": "Конфликт цен, остатков или условий",
            "content": {
              "application/json": {
                "example": {
                  "error": {
                    "code": "CART_CHANGED",
                    "message": "Цены или остатки изменились",
                    "details": {
                      "items": [
                        {
                          "variantId": "00000000-0000-4000-8000-000000000000",
                          "available": 0,
                          "expectedPrice": 12990,
                          "currentPrice": 13990
                        }
                      ]
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "OrderIn": {
        "properties": {
          "address": {
            "$ref": "#/components/schemas/Address"
          },
          "deliveryMethod": {
            "type": "string",
            "const": "manual_delivery",
            "title": "Deliverymethod"
          },
          "customerName": {
            "type": "string",
            "maxLength": 150,
            "minLength": 2,
            "title": "Customername"
          },
          "phone": {
            "type": "string",
            "pattern": "^\\+?[0-9 ()-]{10,25}$",
            "title": "Phone"
          },
          "email": {
            "anyOf": [
              {
                "type": "string",
                "format": "email"
              },
              {
                "type": "null"
              }
            ],
            "title": "Email"
          },
          "paymentMethod": {
            "type": "string",
            "enum": [
              "on_delivery",
              "manual_confirmation"
            ],
            "title": "Paymentmethod"
          },
          "comment": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 2000
              },
              {
                "type": "null"
              }
            ],
            "title": "Comment"
          },
          "personalDataConsent": {
            "type": "boolean",
            "const": true,
            "title": "Personaldataconsent"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "address",
          "deliveryMethod",
          "customerName",
          "phone",
          "paymentMethod",
          "personalDataConsent"
        ],
        "title": "OrderIn"
      },
...
    },
    "securitySchemes": {
      "cartBearer": {
        "type": "http",
        "scheme": "bearer",
        "description": "Токен из POST /api/v1/carts"
      },
      "adminBasic": {
        "type": "http",
        "scheme": "basic",
        "description": "ADMIN_USERNAME и пароль администратора"
      }
    }
  }
}